refactor: Repository 디렉터리 생성 후 이동작업
