+++ src/main/java/subway/domain/Line.java
@@ -1,5 +1,7 @@
 package subway.domain;
 
+import subway.domain.repository.LineRepository;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
